+++ backend/app/models/default_chit_chat.py
@@ -5,7 +5,7 @@ from mojodex_core.entities import *
 from jinja2 import Template
 
 from models.produced_text_manager import ProducedTextManager
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from azure_openai_conf import AzureOpenAIConf
