+++ backend/app/models/documents/document_manager.py
@@ -8,7 +8,7 @@ from mojodex_backend_logger import MojodexBackendLogger
 
 from azure_openai_conf import AzureOpenAIConf
 
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 
 class DocumentManager:
