+++ backend/app/models/home_chat/home_chat_manager.py
@@ -6,7 +6,7 @@ from mojodex_backend_logger import MojodexBackendLogger
 from app import db, placeholder_generator
 from models.knowledge.knowledge_manager import KnowledgeManager
 
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from mojodex_core.entities import *
 from azure_openai_conf import AzureOpenAIConf
 from sqlalchemy import and_
