+++ backend/app/models/produced_text_manager.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from app import db, timing_logger
 import time
 from jinja2 import Template
