+++ backend/app/models/task_classifier.py
@@ -4,7 +4,7 @@ from app import db, log_error, timing_logger
 import os
 from mojodex_core.entities import *
 from jinja2 import Template
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from mojodex_backend_logger import MojodexBackendLogger
 
 from azure_openai_conf import AzureOpenAIConf
