+++ backend/app/models/tasks/task_manager.py
@@ -15,8 +15,8 @@ from models.tasks.task_inputs_manager import TaskInputsManager
 from models.tasks.task_tool_manager import TaskToolManager
 
 from app import socketio_message_sender
-from models.llm_calls.mojodex_openai import MojodexOpenAI
-#from models.llm_calls.mojodex_mistral import MojodexMistralAI, mistral_medium_conf
+from mojodex_core.mojodex_openai import MojodexOpenAI
+#from mojodex_core.mojodex_mistral import MojodexMistralAI, mistral_medium_conf
 from models.produced_text_manager import ProducedTextManager
 from azure_openai_conf import AzureOpenAIConf
 from packaging import version
