+++ backend/app/models/text_edit_action_manager.py
@@ -10,7 +10,7 @@ from models.tasks.task_manager import TaskManager
 from models.tasks.task_executor import TaskExecutor
 from models.produced_text_manager import ProducedTextManager
 
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from azure_openai_conf import AzureOpenAIConf
