+++ backend/app/models/user_audio_file_manager.py
@@ -6,7 +6,7 @@ from models.session import Session as SessionModel
 from mojodex_backend_logger import MojodexBackendLogger
 from azure_openai_conf import AzureOpenAIConf
 
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 
 class UserAudioFileManager:
