+++ backend/app/routes/company.py
@@ -10,10 +10,10 @@ from mojodex_core.entities import *
 from jinja2 import Template
 
 from models.documents.website_parser import WebsiteParser
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 from azure_openai_conf import AzureOpenAIConf
-from models.llm_calls.json_loader import json_decode_retry
+from mojodex_core.json_loader import json_decode_retry
 from app import on_json_error
