+++ backend/app/routes/retrieve_produced_text.py
@@ -6,7 +6,7 @@ from flask_restful import Resource
 from app import db, log_error, document_manager
 from mojodex_core.entities import *
 
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 from azure_openai_conf import AzureOpenAIConf
 from sqlalchemy import func
