+++ backend/app/routes/task_json.py
@@ -5,10 +5,10 @@ from flask import request
 from flask_restful import Resource
 from app import db, log_error
 from mojodex_core.entities import MdTextType
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from azure_openai_conf import AzureOpenAIConf
 from jinja2 import Template
-from models.llm_calls.json_loader import json_decode_retry
+from mojodex_core.json_loader import json_decode_retry
 from app import on_json_error
 
 class TaskJson(Resource):
