+++ backend/app/translator.py
@@ -1,6 +1,6 @@
 from jinja2 import Template
 
-from models.llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 from azure_openai_conf import AzureOpenAIConf
