+++ backend/requirements.txt
@@ -14,7 +14,6 @@ google-search-results==2.4.2
 gevent==23.9.1
 gevent-websocket==0.10.1
 ipython==8.13.0
-psycopg2-binary==2.9.9
 google-auth==2.23.0
 google-auth-oauthlib==1.1.0
 google-auth-httplib2==0.1.1
@@ -32,8 +31,5 @@ grequests==0.7.0
 stripe==6.6.0
 app-store-server-library==1.0.0
 packaging==23.2
-openai==1.3.7
-tiktoken==0.4.0
 boto3==1.28.56
 configparser==6.0.0
-mistralai==0.0.12
