+++ background/app/models/documents/website_parser.py
@@ -7,7 +7,7 @@ from app import document_manager
 
 from background_logger import BackgroundLogger
 from jinja2 import Template
-from llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 from azure_openai_conf import AzureOpenAIConf
