+++ background/app/models/events/todo_daily_emails_generator.py
@@ -5,14 +5,14 @@ from datetime import datetime
 from jinja2 import Template
 
 from background_logger import BackgroundLogger
-from llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 
 from azure_openai_conf import AzureOpenAIConf
 
 from app import send_admin_error_email
 
 from models.events.events_generator import EventsGenerator
-from llm_calls.json_loader import json_decode_retry
+from mojodex_core.json_loader import json_decode_retry
 from app import on_json_error
 
 from models.knowledge.knowledge_collector import KnowledgeCollector
