+++ background/app/models/knowledge/user_knowledge_extractor.py
@@ -2,7 +2,7 @@ import os
 
 import requests
 from jinja2 import Template
-from llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from datetime import datetime
 
 from background_logger import BackgroundLogger
