+++ background/app/models/task_tool_execution/message_writer.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 import requests
 from jinja2 import Template
-from llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from azure_openai_conf import AzureOpenAIConf
 from background_logger import BackgroundLogger
