+++ background/app/models/todos/todos_creator.py
@@ -3,8 +3,8 @@ from datetime import datetime
 
 import requests
 from jinja2 import Template
-from llm_calls.mojodex_openai import MojodexOpenAI
-from llm_calls.json_loader import json_decode_retry
+from mojodex_core.mojodex_openai import MojodexOpenAI
+from mojodex_core.json_loader import json_decode_retry
 from background_logger import BackgroundLogger
 from app import on_json_error
 from azure_openai_conf import AzureOpenAIConf
