+++ background/app/models/user_task_execution_summarizer.py
@@ -1,7 +1,7 @@
 import os
 from datetime import datetime
 import requests
-from llm_calls.mojodex_openai import MojodexOpenAI
+from mojodex_core.mojodex_openai import MojodexOpenAI
 from jinja2 import Template
 
 from background_logger import BackgroundLogger
