+++ background/requirements.txt
@@ -19,7 +19,5 @@ pydantic==1.10.9
 --index-url https://pypi.fury.io/gantry-ml/
 --extra-index-url https://pypi.org/simple/
 gantry-logger==0.0.1a4
-openai==1.3.7
-tiktoken==0.4.0
 boto3==1.28.56
 configparser==6.0.0
