+++ mojodex_core/README.md
@@ -1,6 +1,19 @@
+# Mojodex Core
+
+This directory contains the core functionality of the Mojodex application. The core functionality includes the following:
+
+- ✅ entities
+- 🚧 llm engine
+- 🔜 notifications
+- 🔜 emails / push notifications
+
 # Mojodex Entities
 
-This directory contains the entities that are used in the Mojodex application. An entity is a class that represents a table in the database. The entity class is used to interact with the database table.
+The entities that are used in the Mojodex application. An entity is a class that represents a table in the database. The entity class is used to interact with the database table.
 
 It uses `sqlacodegen` to generate the entity classes from the database schema. The `sqlacodegen` tool is a Python library that can be used to generate SQLAlchemy model code from an existing database.
 
+
+# LLM Engine
+
+All core llm function calls for Mojodex.
