+++ mojodex_core/requirements.txt
@@ -1,2 +1,14 @@
+
+# Database & entities requirements
 pgvector==0.2.4
 sqlalchemy==1.4.31
+psycopg2-binary==2.9.9
+
+# LLM requirements
+mistralai==0.0.12
+openai==1.3.7
+tiktoken==0.4.0
+
+
+# Other requirements
+pydub==0.25.1
